django/requirements.txt
Django==2.0.5
django-filter==1.1.0
djangorestframework==3.7.3
factory-boy==2.10.0
feedparser==5.2.1
psycopg2==2.7.3.2
pytest-django==3.1.2
uWSGI==2.0.17
